LayoutTests/imported/w3c/web-platform-tests/mimesniff/sniffing/html.window-expected.txt


PASS HTML is not sniffed for a "feed": atom
PASS HTML is not sniffed for a "feed": rss
